# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VipGoPay API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  cat?: string;
  price: number;
  stock: number;
  cond?: string;
  specs?: string;
}

export interface ProductInput {
  name: string;
  cat?: string;
  price: number;
  stock?: number;
  cond?: string;
  specs?: string;
}

export interface StockUpdate {
  stock: number;
}

export interface CustomerInfo {
  name: string;
  phone: string;
  wilaya: string;
  commune: string;
}

export interface Order {
  id: string;
  cust: CustomerInfo;
  product: string;
  productId: number;
  price: number;
  status: string;
  time: string;
}

export interface OrderInput {
  id: string;
  cust: CustomerInfo;
  product: string;
  productId: number;
  price: number;
  status: string;
  time: string;
}

export interface OrderStatusUpdate {
  status: string;
}

export interface WilayaRate {
  wilaya: string;
  office: string;
  home: string;
}

export interface StoreSettings {
  shopName?: string;
  shopType?: string;
  location?: string;
  hours?: string;
  officeDesk?: string;
  officeHome?: string;
  extra?: string;
  model?: string;
  email?: string;
  muteNotif?: boolean;
  wilayaRates?: WilayaRate[];
}

export interface ChatMessage {
  role: string;
  content: string;
}

export interface ChatRequest {
  model?: string;
  messages: ChatMessage[];
  storeConfig?: StoreSettings;
  productsList?: Product[];
  customerInfo?: CustomerInfo;
}

export interface ChatReply {
  reply: string;
}

export interface NotifyResult {
  channel: string;
  ok: boolean;
  /** @nullable */
  error?: string | null;
}

export type NotifyRequestIntegrations = { [key: string]: unknown };

export interface NotifyRequest {
  order: Order;
  integrations?: NotifyRequestIntegrations;
  /** @nullable */
  customerEmail?: string | null;
}

export interface NotifyResponse {
  results: NotifyResult[];
  /** @nullable */
  warning?: string | null;
}

export interface OtpRequest {
  email: string;
}

export interface OtpRequestResult {
  success: boolean;
  /** @nullable */
  devCode?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface OtpVerify {
  email: string;
  code: string;
}

export interface OtpVerifyResult {
  success: boolean;
  /** @nullable */
  error?: string | null;
}