# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VipGoPay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';
import type { CustomerInfo } from './customerInfo';
import type { Product } from './product';
import type { StoreSettings } from './storeSettings';

export interface ChatRequest {
  model?: string;
  messages: ChatMessage[];
  storeConfig?: StoreSettings;
  productsList?: Product[];
  customerInfo?: CustomerInfo;
}