# lib/api-zod/src/generated/types/customerInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VipGoPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerInfo {
  name: string;
  phone: string;
  wilaya: string;
  commune: string;
}