# lib/api-zod/src/generated/types/notifyRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VipGoPay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotifyRequestIntegrations } from './notifyRequestIntegrations';
import type { Order } from './order';

export interface NotifyRequest {
  order: Order;
  integrations?: NotifyRequestIntegrations;
  /** @nullable */
  customerEmail?: string | null;
}