# lib/api-zod/src/generated/types/notifyResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VipGoPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NotifyResult {
  channel: string;
  ok: boolean;
  /** @nullable */
  error?: string | null;
}