# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VipGoPay API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerInfo } from './customerInfo';

export interface Order {
  id: string;
  cust: CustomerInfo;
  product: string;
  productId: number;
  price: number;
  status: string;
  time: string;
}