# lib/api-zod/src/generated/types/otpRequestResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VipGoPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OtpRequestResult {
  success: boolean;
  /** @nullable */
  devCode?: string | null;
  /** @nullable */
  error?: string | null;
}