# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VipGoPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  cat?: string;
  price: number;
  stock?: number;
  cond?: string;
  specs?: string;
}