# lib/api-zod/src/generated/types/wilayaRate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VipGoPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WilayaRate {
  wilaya: string;
  office: string;
  home: string;
}